Re-organized, and wrote documentation for almost every piece of code
