+++ src/ksm/sections/mod.rs
@@ -1,3 +1,4 @@
+//! A collection of the sections that can be contained within a KSM file
 mod argument_section;
 mod code_section;
 mod debug_section;
