--- src/ksmfile/errors.rs
@@ -1,4 +1,0 @@
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum KSMReadError {}
